+++ gulpfile.js
@@ -24,7 +24,7 @@ import { webpCompress, genSvgSprite, copyImages } from './.gulp/images.js';
 import { openServer, openBrowser, bumper, cleanBuild, cleanDist, cleanHTML, openProxyTunnel, bs } from './.gulp/server.js';
 
 // Misc
-import { copyTest, copyFiles, copyBuildFiles, copyVideo, copyFonts, copyIcons } from './.gulp/misc.js';
+import { copyTest, copyFiles, copyBuildFiles, copyVideo, copyFonts, copyIcons, copyReport } from './.gulp/misc.js';
 
 // Tests
 import { mobileTestRes, htmlSpeedRes, cssTestRes } from './.gulp/tests.js';
@@ -43,8 +43,13 @@ const { parallel, series, watch } = gulp;
  * Start
  */
 
+
+
 // Disable deprecation warnings
 process.noDeprecation = true;
+
+// Increase event listener limit
+process.setMaxListeners(0);
 
 // Clear shell screen
 console.clear();
@@ -71,16 +76,23 @@ const concatJsLibs = (done) => jsConcatVendorLibs(jsVendorList, done);
 /**
  * Watcher
  */
+
+const cfgWatch = {
+  usePolling: true,
+  interval: 500,
+  ignoreInitial: true
+};
+
 const watcher = (done) => {
-  const watchOptions = { usePolling: true, interval: 500 };
-  watch('./src/scss/**/*.scss', watchOptions, series(scss, stylesReload));
-  watch('./src/**/*.html', watchOptions, series(htmlGenerate, cleanHTML, htmlReload, testHtml));
-  watch('./src/javascript/**/*.js', watchOptions, series(series(roll, concatJsLibs), scriptsReload));
-  watch('./src/images/**/*', watchOptions, series(webpCompress, copyImages));
-  watch('./src/favicons/**/*', watchOptions, series(copyIcons));
-  watch('./src/fonts/**/*', watchOptions, series(copyFonts));
-  watch('./src/video/**/*', watchOptions, series(copyVideo));
-  watch('./src/test/**/*', watchOptions, series(copyTest));
+  watch('./src/scss/**/*.scss', cfgWatch, series(scss, stylesReload));
+  watch('./src/**/*.html', cfgWatch, series(htmlGenerate, cleanHTML, htmlReload, testHtml));
+  watch('./src/javascript/**/*.js', cfgWatch, series(series(roll, concatJsLibs), scriptsReload));
+  watch('./src/images/**/*', cfgWatch, series(webpCompress, copyImages));
+  watch('./src/favicons/**/*', cfgWatch, series(copyIcons));
+  watch('./src/fonts/**/*', cfgWatch, series(copyFonts));
+  watch('./src/video/**/*', cfgWatch, series(copyVideo));
+  watch('./src/test/**/*', cfgWatch, series(copyTest));
+  watch('./src/report/**/*', cfgWatch, series(copyReport));
   done();
 };